Renamed the dto and di folders to more descriptive names. Added comments to the database classes. Removed unnecessary files.
