+++ app/src/main/java/com/example/pricingpal/model/datatransferobjects/CategoryDTO.kt
@@ -1,4 +1,4 @@
-package com.example.pricingpal.model.dto
+package com.example.pricingpal.model.datatransferobjects
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
