+++ app/src/main/java/com/example/pricingpal/model/repositories/impl/CategoryRepositoryImpl.kt
@@ -1,30 +1,37 @@
 package com.example.pricingpal.model.repositories.impl
 
 import com.example.pricingpal.model.Category
-import com.example.pricingpal.model.dto.CategoryDTO
+import com.example.pricingpal.model.datatransferobjects.CategoryDTO
 import com.example.pricingpal.model.repositories.CategoryRepository
 import io.github.jan.supabase.postgrest.Postgrest
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 import javax.inject.Inject
 
+/**
+ * Class: CategoryRepositoryImpl
+ * @author Connor Murdock
+ * @version 1
+ * @written 11/20/2023
+ * This class acts as an implemented version of the CategoryRepository Interface the app can use to interact with the database's Categories
+ */
 class CategoryRepositoryImpl @Inject constructor (
     private val postgrest: Postgrest
-    // private val storage: Storage,
     ) : CategoryRepository {
-    //Create or get a list of all Categories or a single Category
+    //This function will create a new category and push it to the database using the given Category object
     override suspend fun createCategory(category: Category): Boolean {
         TODO("Not yet implemented")
     }
 
+    //Gets a list of all categories from the database
     override suspend fun getCategories(): List<CategoryDTO>? {
         return withContext(Dispatchers.IO) {
             val results = postgrest["category"].select().decodeList<CategoryDTO>()
             results
         }
-
     }
 
+    //Gets a specific category by its id value
     override suspend fun getCategory(id: Int): CategoryDTO {
         return withContext(Dispatchers.IO) {
             postgrest["category"].select {
@@ -33,12 +40,13 @@ class CategoryRepositoryImpl @Inject constructor (
         }
     }
 
-    // Delete a Category
+    //This function will remove a category from the database
     override suspend fun deleteCategory(id: Int) {
         TODO("Not yet implemented")
+        //This function will remove a category from the database
     }
 
-    // Update a Category
+    //This function will update an existing category's name using the given String value that will be input by the user
     override suspend fun updateCategory(name: String) {
         TODO("Not yet implemented")
     }
