+++ app/src/main/java/com/example/pricingpal/viewmodel/CategoryViewModel.kt
@@ -1,12 +1,11 @@
 package com.example.pricingpal.viewmodel
 
-import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.pricingpal.model.Category
 import com.example.pricingpal.model.Item
-import com.example.pricingpal.model.dto.CategoryDTO
-import com.example.pricingpal.model.dto.ItemDTO
+import com.example.pricingpal.model.datatransferobjects.CategoryDTO
+import com.example.pricingpal.model.datatransferobjects.ItemDTO
 import com.example.pricingpal.model.repositories.CategoryRepository
 import com.example.pricingpal.model.repositories.ItemRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
@@ -15,48 +14,60 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
+/**
+ * Class: CategoryViewModel
+ * @author Connor Murdock
+ * @version 1
+ * @written 11/20/2023
+ * This class is the ViewModel for the Category/Item information. When this ViewModel is created it will connect
+ * to the SupaBase database, pull the list of Categories and Items, then turn them into usable data for the app.
+ */
 @HiltViewModel
 class CategoryViewModel @Inject constructor(
     private val categoryRepository: CategoryRepository,
     private val itemRepository: ItemRepository
 ) : ViewModel() {
-
-
+    //A hashmap of the categories, where the key is the Category name as a String, and the value is the corresponding Category object
+    //This hashmap is used by the navigation to know which category was selected when a category is tapped on the main screen
     val categories = HashMap<String, Category>()
 
+    //lists of the categories from the database that are updated by the ViewModel
     private val _categoryList = MutableStateFlow<List<Category>?>(listOf())
     val categoryList: Flow<List<Category>?> = _categoryList
 
+    //When this ViewModel is created, call these two functions
     init {
         getCategories()
         observeCategories()
     }
 
+    //Observes the categoryList and calls the populateCategories function once the data has been retrieved
     fun observeCategories() {
         viewModelScope.launch {
             categoryList.collect { categoryList ->
                 categoryList?.let {
                     populateCategories(it)
                 }
-
             }
         }
     }
 
+    //Populates the HashMap with the Category objects from the database
     fun populateCategories(categoryList: List<Category>) {
         for (c in categoryList) {
             categories.put(c.category, c)
-            Log.e("Category", c.toString())
         }
     }
 
-    fun getCategories(){
-                viewModelScope.launch {
-                    val categories2 = categoryRepository.getCategories()
-                    _categoryList.emit(categories2?.map { it -> it.asDomainModel() })
-                }
-}
+    //Gets the list of categories from the database and emits it to _categoryList
+    fun getCategories() {
+        viewModelScope.launch {
+            val categories2 = categoryRepository.getCategories()
+            _categoryList.emit(categories2?.map { it -> it.asDomainModel() })
+        }
+    }
 
+    //When the Category data from the database is pulled, this function will translate it from a DTO to the real object
     suspend fun CategoryDTO.asDomainModel(): Category {
         return Category(
             category = this.categoryName,
@@ -64,10 +75,12 @@ class CategoryViewModel @Inject constructor(
         )
     }
 
+    //Gets the list of item from the database and returns it as an ArrayList of Items
     suspend fun getItems(categoryId: Int): ArrayList<Item>? {
         return itemRepository.getItems(categoryId)?.map { it.asDomainModel() } as ArrayList<Item>?
     }
 
+    //When the Item data from the database is pulled, this function will translate it from a DTO to the real object
     private fun ItemDTO.asDomainModel(): Item {
         return Item(
             id = this.itemId,
@@ -75,7 +88,6 @@ class CategoryViewModel @Inject constructor(
             price = this.price.toDouble(),
             categoryId = this.categoryId,
             organizationId = this.organizationId.toString()
-
         )
     }
 }
